requirements.txt
streamlit
openai-whisper
torch
audio-recorder-streamlit
moviepy
googletrans==4.0.0-rc1